transformations/requirements.txt
pandas
minio
psycopg2-binary
